Combine common GTK timer code.
