+++ lib/matplotlib/backends/_backend_gtk.py
@@ -7,12 +7,12 @@ import logging
 import matplotlib as mpl
 from matplotlib import cbook
 from matplotlib.backend_bases import (
-    _Backend,
+    _Backend, TimerBase,
 )
 
 # The GTK3/GTK4 backends will have already called `gi.require_version` to set
 # the desired GTK.
-from gi.repository import Gio, Gtk
+from gi.repository import Gio, GLib, Gtk
 
 
 _log = logging.getLogger(__name__)
@@ -60,6 +60,42 @@ def _create_application():
             _application = app
 
     return _application
+
+
+class TimerGTK(TimerBase):
+    """Subclass of `.TimerBase` using GTK timer events."""
+
+    def __init__(self, *args, **kwargs):
+        self._timer = None
+        super().__init__(*args, **kwargs)
+
+    def _timer_start(self):
+        # Need to stop it, otherwise we potentially leak a timer id that will
+        # never be stopped.
+        self._timer_stop()
+        self._timer = GLib.timeout_add(self._interval, self._on_timer)
+
+    def _timer_stop(self):
+        if self._timer is not None:
+            GLib.source_remove(self._timer)
+            self._timer = None
+
+    def _timer_set_interval(self):
+        # Only stop and restart it if the timer has already been started.
+        if self._timer is not None:
+            self._timer_stop()
+            self._timer_start()
+
+    def _on_timer(self):
+        super()._on_timer()
+
+        # Gtk timeout_add() requires that the callback returns True if it
+        # is to be called again.
+        if self.callbacks and not self._single:
+            return True
+        else:
+            self._timer = None
+            return False
 
 
 class _BackendGTK(_Backend):
