+++ lib/matplotlib/backends/backend_gtk4.py
@@ -32,7 +32,9 @@ except ValueError as e:
 from gi.repository import Gio, GLib, GObject, Gtk, Gdk, GdkPixbuf
 from ._backend_gtk import (
     _create_application, _shutdown_application,
-    backend_version, _BackendGTK)
+    backend_version, _BackendGTK,
+    TimerGTK as TimerGTK4,
+)
 
 
 _log = logging.getLogger(__name__)
@@ -48,42 +50,6 @@ def _mpl_to_gtk_cursor(mpl_cursor):
         Cursors.RESIZE_HORIZONTAL: "ew-resize",
         Cursors.RESIZE_VERTICAL: "ns-resize",
     }, cursor=mpl_cursor)
-
-
-class TimerGTK4(TimerBase):
-    """Subclass of `.TimerBase` using GTK4 timer events."""
-
-    def __init__(self, *args, **kwargs):
-        self._timer = None
-        super().__init__(*args, **kwargs)
-
-    def _timer_start(self):
-        # Need to stop it, otherwise we potentially leak a timer id that will
-        # never be stopped.
-        self._timer_stop()
-        self._timer = GLib.timeout_add(self._interval, self._on_timer)
-
-    def _timer_stop(self):
-        if self._timer is not None:
-            GLib.source_remove(self._timer)
-            self._timer = None
-
-    def _timer_set_interval(self):
-        # Only stop and restart it if the timer has already been started
-        if self._timer is not None:
-            self._timer_stop()
-            self._timer_start()
-
-    def _on_timer(self):
-        super()._on_timer()
-
-        # Gtk timeout_add() requires that the callback returns True if it
-        # is to be called again.
-        if self.callbacks and not self._single:
-            return True
-        else:
-            self._timer = None
-            return False
 
 
 class FigureCanvasGTK4(Gtk.DrawingArea, FigureCanvasBase):
